Some old fixes + ArchiCAD 20 support
The addon and its respective C++ code is now available. However, the
plug-in for ArchiCAD 20 cannot delete stories, meaning that the function
delete-levels does not work. I was not able to find a solution to this
problem. The stories' IDs seem to be correct, as I'm able to create
stories. Maybe a problem with the API?